+++ COMP313003_202101_Project/Areas/Accounting/Controllers/Api/InvoiceTypeController.cs
@@ -57,7 +57,7 @@ namespace COMP313003_202101_Project.Controllers.Api
         public IActionResult Remove([FromBody]CrudViewModel<InvoiceType> payload)
         {
             InvoiceType invoiceType = _context.InvoiceType
-                .Where(x => x.InvoiceTypeId == (int)payload.key)
+                .Where(x => x.InvoiceTypeId == (int)(long)payload.key)
                 .FirstOrDefault();
             _context.InvoiceType.Remove(invoiceType);
             _context.SaveChanges();
