+++ COMP313003_202101_Project/Areas/Inventory/Controllers/Api/UnitOfMeasureController.cs
@@ -56,7 +56,7 @@ namespace COMP313003_202101_Project.Areas.Inventory.Controllers.Api
         public IActionResult Remove([FromBody]CrudViewModel<UnitOfMeasure> payload)
         {
             UnitOfMeasure unitOfMeasure = _context.UnitOfMeasure
-                .Where(x => x.UnitOfMeasureId == (int)payload.key)
+                .Where(x => x.UnitOfMeasureId == (int)(long)payload.key)
                 .FirstOrDefault();
             _context.UnitOfMeasure.Remove(unitOfMeasure);
             _context.SaveChanges();
